requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
antlr4-python3-runtime==4.9.3
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
docker-pycreds==0.4.0
filelock==3.16.1; python_version >= '3.8'
fsspec==2024.12.0; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
jinja2==3.1.4; python_version >= '3.7'
lightning-utilities==0.11.9; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.4.2; python_version >= '3.10'
numpy==2.2.0; python_version >= '3.10'
omegaconf==2.3.0; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.0.0; python_version >= '3.9'
platformdirs==4.3.6; python_version >= '3.8'
protobuf==5.29.2; sys_platform != 'linux'
psutil==6.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.10.4; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
sentry-sdk==2.19.2; python_version >= '3.6'
setproctitle==1.3.4; python_version >= '3.8'
setuptools==75.6.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1; python_version >= '3.7'
sympy==1.13.1; python_version >= '3.9'
torch==2.5.1; python_full_version >= '3.8.0'
torchmetrics==1.6.0; python_version >= '3.9'
torchvision==0.20.1; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
wandb==0.19.1; python_version >= '3.8'
